restructuring of functions in separate files
